Refactor: move pure Markdown logic into com.editora.markdown
Extract the 11 toolkit-free Markdown cores (editing: MarkdownEdit/Inline/Lines
/Heading/Table; lint: MarkdownLint/LintConfig/LintDirectives/LintFix; plus
MarkdownOutline + MarkdownImagePaste) and their 10 unit tests out of
com.editora.editor into a new com.editora.markdown package (editor: 78 -> 67
files). The FX MarkdownRenderer, preview/lint overlays + stripes,
MarkdownFormatBar, and the image loaders stay in editor/ as the editor's
rendering surfaces.

git mv preserves history; consumer references rewired with a boundary-safe
rewrite so the stayers MarkdownLintService/Overlay/Stripe are untouched.
MarkdownLintDirectives (package-private) + its test move together to keep
package access. Internal package: no exports/opens/module-info change, no
behavior/dependency change.
