+++ CHANGELOG.md
@@ -73,6 +73,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - **A "Compare with HEAD" diff no longer blanks its HEAD pane on refresh.** In a window with no project open, opening a file's vs-HEAD diff and then triggering a refresh (clicking an "apply change" chevron, or regaining window focus) could empty the left (HEAD) pane and render the whole working side as all-added. The diff's HEAD side was re-fetched against the *live* repo root, which the Git status machine drops to "none" when the active tab is a diff tab with no backing buffer (a No-Project window has no project root to fall back to) — so the re-fetch ran against a null root and returned nothing. The diff now captures the repo root when it's opened and reuses it on every refresh. Affected all Git-backed diffs (vs HEAD, vs commit, Git-panel staged/unstaged rows, a commit's file, and the Local File History revision diff). Covered by a new FX regression test.
 
 ### Changed
+- **Refactor: pure Markdown logic moved to a `com.editora.markdown` package.** The 11 toolkit-free Markdown cores (editing, lint, outline, image-paste) and their tests moved out of `com.editora.editor` into their own package; the FX renderer and editor overlays stay. Internal-only — no behavior, dependency, or `module-info` change.
 - **The "New File From Template" picker is bigger.** Widened (and given a taller minimum) so the longer template descriptions are no longer clipped (the default width showed a horizontal scrollbar).
 - **Personal Notes show an inline marker at the note's start instead of a gutter icon.** A small amber corner-triangle now marks where a note begins (the selected character/word/range), reinforcing that notes are anchored to the selection — not the line. The gutter note slot is gone; edit the note at the caret with the new `notes.editNote` command (the inline highlight + hover tooltip + Notes tool window are unchanged).
 - **The Undo History tool-window stripe button is now off by default.** With the new Undo History popup (`undoHistory.jump`) as the primary entry point, the side-stripe icon is hidden until you enable it (Settings → Tool Windows, or the `tool.undoHistory` command). Existing sessions keep whatever visibility you had.
