+++ src/main/java/com/editora/editor/EditorBuffer.java
@@ -42,6 +42,12 @@ import com.editora.completion.CompletionEngine;
 import com.editora.completion.CompletionProvider;
 import com.editora.logviewer.LogFilter;
 import com.editora.logviewer.LogLevel;
+import com.editora.markdown.MarkdownEdit;
+import com.editora.markdown.MarkdownHeading;
+import com.editora.markdown.MarkdownInline;
+import com.editora.markdown.MarkdownLines;
+import com.editora.markdown.MarkdownLint;
+import com.editora.markdown.MarkdownTable;
 import com.editora.snippet.ParsedSnippet;
 import com.editora.snippet.Snippet;
 import com.editora.snippet.SnippetParser;
