+++ src/main/java/com/editora/markdown/MarkdownLintDirectives.java
@@ -1,4 +1,4 @@
-package com.editora.editor;
+package com.editora.markdown;
 
 import java.util.HashSet;
 import java.util.Locale;
