+++ src/test/java/com/editora/markdown/MarkdownLintFixTest.java
@@ -1,4 +1,4 @@
-package com.editora.editor;
+package com.editora.markdown;
 
 import java.util.Set;
